+++ PiedraPapelTijera/PiedraPapelTijera.html
@@ -1,0 +1,91 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>🪨 📜 ✂️</title>
+    <!-- SWEETALERT2 -->
+    <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
+    <!-- LOGO -->
+    <link rel="shortcut icon" href="./assets/logo.png" type="image/x-icon">
+    <!-- BOOTSTRAP -->
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.13.1/font/bootstrap-icons.min.css">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-LN+7fdVzj6u52u30Kp6M/trliBMCMKTyK833zpbD+pXdCLuTusPj697FH4R/5mcr" crossorigin="anonymous">
+    <!-- CSS -->
+    <link rel="stylesheet" href="./style/index.css">
+    <link rel="stylesheet" href="./style/header.css">
+    <link rel="stylesheet" href="./style/main.css">
+    <link rel="stylesheet" href="./style/footer.css">
+    <link rel="stylesheet" href="./style/script.css">
+</head>
+
+<body>
+    <header>
+        <div class="container-header">
+            <div class="points d-flex justify-content-center align-items-center gap-3">
+                <div class="user d-flex flex-column justify-content-center align-items-center">
+                    <i class="bi bi-person"></i>
+                    <p class="points-user m-1">0</p>
+                </div>
+                <div class="computer d-flex flex-column justify-content-center align-items-center">
+                    <i class="bi bi-pc-display"></i>
+                    <p class="points-computer m-1">0</p>
+                </div>
+            </div>
+            <div class="title text-center">
+                <h1 class="first-title">🪨 📜 ✂️</h1>
+            </div>
+        </div>
+    </header>
+
+    <main>
+        <div class="container-fluid">
+            <div class="row">
+                <div class="text-center">
+                    <h2 class="second-title">PIEDRA PAPEL TIJERA</h2>
+                    <p class="description">QUIERES JUGAR ¡PRESIONA EL BOTON!</p>
+                    <button class="btn btn-play">JUGAR</button>
+                </div>
+                <div class="game d-flex flex-column justify-content-center align-items-center">
+                    <div class="options-img d-flex justify-content-center align-items-center gap-3">
+                        <!-- IMAGENES/OPCIONES EN JS -->
+                    </div>
+                    <div class="instructions text-center mt-3">
+                        <h2 class="description">INSTRUCCIONES</h2>
+                        <p class="description-instructions">
+                            ELIGE UNA OPCION DE LAS IMAGENES DE ARRIBA
+                        </p>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </main>
+
+    <footer>
+        <div class="container-footer">
+            <div class="container-redes d-flex justify-content-start align-items-end gap-5">
+                <a href="https://github.com/LautaroLeall" target="_blank" class="footer-link">
+                    <i class="bi bi-github"></i>
+                </a>
+                <a href="https://www.instagram.com/lautaro_leall/" target="_blank" class="footer-link">
+                    <i class="bi bi-instagram"></i>
+                </a>
+                <a href="https://www.linkedin.com/in/lauldp/" target="_blank" class="footer-link">
+                    <i class="bi bi-linkedin"></i>
+                </a>
+            </div>
+        </div>
+    </footer>
+
+    <!-- BOOTSTRAP JS -->
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-ndDqU0Gzau9qJ1lfW4pNLlhNTkCfHzAVBReH9diLvGRem5+R9g2FzA8ZGN954O5Q"
+        crossorigin="anonymous"></script>
+
+    <!-- SCRIPT PRINCIPAL -->
+    <script type="module" src="./scripts/main.js"></script>
+</body>
+
+</html>
