+++ PiedraPapelTijera/style/footer.css
@@ -1,0 +1,24 @@
+/* Footer */
+.container-footer {
+    width: 60vw;
+}
+
+.container-redes {
+    font-size: 1.1rem;
+    transition: all 0.3s ease;
+}
+
+.footer-link {
+    padding: 0.3rem;
+    transition: all 0.3s ease;
+    border-radius: 50%;
+    color: rgb(30, 80, 80);
+    background-color: rgba(70, 100, 100, 0.6);
+    border: 5px solid rgb(70, 100, 100);
+}
+
+.footer-link:hover {
+    transition: all 0.3s ease;
+    transform: scale(0.9);
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+}
