+++ PiedraPapelTijera/style/main.css
@@ -1,0 +1,44 @@
+/* Main */
+.container-fluid {
+    width: 100%;
+    height: 50vh;
+}
+
+/* Títulos */
+.first-title {
+    font-size: 7rem;
+    text-shadow: 5px 5px 20px rgb(70, 150, 150);
+}
+
+.second-title {
+    font-size: 2rem;
+    letter-spacing: 0.5rem;
+    text-shadow: 5px 5px 10px rgb(70, 150, 150);
+    color: rgb(30, 80, 80);
+}
+
+.description,
+.description-instructions {
+    font-size: 1rem;
+    letter-spacing: 0.5rem;
+    font-weight: 600;
+    color: rgb(30, 80, 80);
+}
+
+/* Botónes */
+.btn-play {
+    letter-spacing: 0.3rem;
+    font-weight: 600;
+    background-color: rgba(70, 100, 100, 0.6);
+    border: 5px solid rgb(70, 100, 100);
+    color: rgb(30, 80, 80);
+    transition: all 0.3s ease-in-out;
+    padding: 0.5rem 2rem;
+    border-radius: 12px;
+}
+
+.btn-play:hover {
+    background-color: rgb(30, 80, 80);
+    color: rgba(70, 100, 100, 0.6);
+    border: 5px solid rgb(70, 100, 100);
+}
