+++ PiedraPapelTijera/style/script.css
@@ -1,0 +1,68 @@
+/* Botón Reset/Salir (heredan estilos del botón JUGAR) */
+.btn-reset,
+.btn-exit {
+    margin: 0.5rem 0;
+}
+
+/* Imágenes de opciones */
+.opcion-img {
+    width: 10rem;
+    border-radius: 50%;
+    padding: 1rem;
+    background-color: rgba(70, 100, 100, 0.6);
+    border: 5px solid rgb(70, 100, 100);
+    transition: all 0.3s ease;
+    cursor: pointer;
+}
+
+/* Hover IMG */
+.opcion-img:hover {
+    transform: scale(0.9) rotateZ(-5deg);
+    background-color: rgb(30, 80, 80);
+    border-color: rgb(70, 100, 100);
+    box-shadow: 0 0 20px rgba(0, 0, 0, 0.5);
+}
+
+/* Animaciones base */
+@keyframes fadeIn {
+    0% {
+        opacity: 0;
+        transform: scale(0.5);
+    }
+
+    100% {
+        opacity: 1;
+        transform: scale(1);
+    }
+}
+
+@keyframes shake {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    25% {
+        transform: rotate(-10deg);
+    }
+
+    50% {
+        transform: rotate(10deg);
+    }
+
+    75% {
+        transform: rotate(-10deg);
+    }
+
+    100% {
+        transform: rotate(0deg);
+    }
+}
+
+/* Clases animación */
+.animate-fade {
+    animation: fadeIn 0.5s ease forwards;
+}
+
+.animate-shake {
+    animation: shake 0.5s ease;
+}
